% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stellenbosch_colours.R
\docType{data}
\name{stellenbosch_colours}
\alias{stellenbosch_colours}
\title{ggplot colour palettes inspired by the colours of Stellenbosch and the Western Cape}
\format{An object of class \code{character} of length 5.}
\usage{
stellenbosch_colours
}
\description{
ggplot colour palettes inspired by the colours of Stellenbosch and the Western Cape
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
 geom_point(size = 4) +
 scale_colour_stellenbosch()

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
 geom_point(size = 6, alpha = 0.8) +
 scale_colour_stellenbosch(discrete = FALSE)

ggplot(iris, aes(Species,Sepal.Length,fill=Species)) +
  geom_bar(stat="identity",alpha=.8) +
  scale_fill_stellenbosch()

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point(size = 4) +
  scale_colour_stellenbosch("wine") +
  theme_dark()

ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
  geom_point(size = 6, alpha = 0.8) +
  scale_colour_stellenbosch(discrete = FALSE, palette = "wine")

ggplot(iris, aes(Species,Sepal.Length,fill=Species)) +
  geom_bar(stat="identity",alpha=.8) +
  scale_fill_stellenbosch("wine")

ggplot(diamonds,aes(x=carat,y=price, colour= price)) +
geom_point(alpha=.3) +
scale_colour_stellenbosch(discrete=F,"main")

ggplot(diamonds,aes(x=carat,y=price, colour= price)) +
geom_point(alpha=.3) +
scale_colour_stellenbosch(discrete=F,"wine")

colorblindness test
pal.safe(stellenbosch_pal())
pal.safe(stellenbosch_pal("wine"))

quality test
pal.test(stellenbosch_pal())
pal.test(stellenbosch_pal("wine"))

bands
pal.bands(stellenbosch_pal(),stellenbosch_pal("wine"))
pal.bands(stellenbosch_pal(),stellenbosch_pal("wine"),main = "Discrete",show.names=T)
}
\keyword{datasets}
